third_party/WebKit/LayoutTests/scrollbars/destroy-scrollbar-if-scrollsizes-lesser-than-contentsizes.html

<!DOCTYPE HTML>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/run-after-layout-and-paint.js"></script>
<div id="scroller" style="overflow:auto; width:100px; height:100px">
  <div id="content" style="width:100px; height:100px; background:blue">
  </div>
</div>
<script>
async_test(function(t) {
    var scroller = document.getElementById("scroller");
    var content = document.getElementById("content");
    content.style.height = "110px";
    runAfterLayoutAndPaint(t.step_func(function() {
        assert_equals(scroller.clientWidth, 85);
        content.style.height = "100px";
        runAfterLayoutAndPaint(t.step_func_done(function() {
            assert_equals(scroller.clientHeight, 100);
        }));
    }));
});
</script>
